Pull request refactorization.
