+++ web/WEB-INF/layout.xhtml
@@ -39,6 +39,7 @@
                         <li><a href="#{request.contextPath}/jobs/showJobs.xhtml">Jobs</a></li>
                         <li><a href="#{request.contextPath}/features/showAllFeatures.xhtml">Features</a></li>
                         <li><a href="#{request.contextPath}/users/showUsers.xhtml">Users</a></li>
+                        <li><a href="#{request.contextPath}/statistics/statistics.xhtml">Statistics</a></li>
                     </ul>
                     
                     <c:if test="#{userController.getCurrentUser() != null}">
